Adding repository junit tests

+++ src/main/java/com/example/rewardssvc/model/Rewards.java
@@ -10,5 +10,5 @@ import java.util.List;
 public class Rewards {
 
     private List<RewardsEntry> monthlyRewards;
-    private Long totalRewardPoints;
+    private Double totalRewardPoints;
 }
